Fixed bag to suit specs
Changed it so it no longer relies on the array. It now functions as a two linked lists.

I built a test using a basic DElinked list (not the one we currently have) just for testing and it worked perfectly!

It is currently 5:11am I will fix what is left as to work with our linked list tomorrow.